+++ src/main/java/com/sustech/gamercenter/dao/GameCommentRepository.java
@@ -12,10 +12,10 @@ import java.util.List;
 @Repository
 public interface GameCommentRepository extends JpaRepository<GameComment, Long> {
     @Query(value = "select * from game_comment where user_id = :UID and visible = True", nativeQuery = true)
-    List<GameComment> getCommentByUser(int UID);
+    List<GameComment> getCommentByUser(long UID);
 
     @Query(value = "select * from game_comment where game_id = :GID and visible = True", nativeQuery = true)
-    List<GameComment> getCommentByGame(int GID);
+    List<GameComment> getCommentByGame(long GID);
 
     @Query(value = "select * from game_comment where visible = True", nativeQuery = true)
     List<GameComment> getAllComment();
@@ -23,10 +23,10 @@ public interface GameCommentRepository extends JpaRepository<GameComment, Long> 
     @Transactional
     @Modifying
     @Query(value = "update game_comment set visible = False where game_comment_id = :game_comment_id", nativeQuery = true)
-    int  deleteComment(int game_comment_id);
+    int  deleteComment(long game_comment_id);
 
     @Transactional
     @Modifying
     @Query(value = "insert into game_comment(user_id, game_id, content, grade) value(?1,?2,?3,?4)",nativeQuery = true)
-    int addComment(int UID, int GID, String content, double grade);
+    int addComment(long UID, long GID, String content, double grade);
 }
